[fix] warnings

+++ util/CircularBuff.h
@@ -23,7 +23,7 @@ Linker_require("util/CircularBuff.c")
 typedef struct uv_buff_circular {
     uv_buf_t *buffs; // array of buffers
     size_t max_size; // number of elements in buffs
-    int size; // current size
+    size_t size; // current size
     // private
     uv_buf_t *current_element; // last element
 } uv_buff_circular;
